refact(*): unify server and workflow common package
