+++ cmd/workflow/coordinator/main.go
@@ -9,7 +9,7 @@ import (
 	log "github.com/sirupsen/logrus"
 
 	"github.com/caicloud/cyclone/pkg/common"
-	k8sclient "github.com/caicloud/cyclone/pkg/common"
+	utilk8s "github.com/caicloud/cyclone/pkg/util/k8s"
 	"github.com/caicloud/cyclone/pkg/workflow/coordinator"
 )
 
@@ -25,7 +25,7 @@ func main() {
 	var message string
 
 	// Create k8s clientset and registry system signals for exit.
-	client, err := k8sclient.GetClient("", *kubeConfigPath)
+	client, err := utilk8s.GetClient(*kubeConfigPath)
 	if err != nil {
 		log.Errorf("Get k8s client error: %v", err)
 		os.Exit(1)
