+++ pkg/server/handler/v1alpha1/template.go
@@ -10,11 +10,10 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 
 	"github.com/caicloud/cyclone/pkg/apis/cyclone/v1alpha1"
-	"github.com/caicloud/cyclone/pkg/server/common"
+	"github.com/caicloud/cyclone/pkg/common"
 	"github.com/caicloud/cyclone/pkg/server/handler"
 	"github.com/caicloud/cyclone/pkg/server/types"
 	"github.com/caicloud/cyclone/pkg/util/cerr"
-	wfcommon "github.com/caicloud/cyclone/pkg/workflow/common"
 )
 
 // ListTemplates get templates the given tenant has access to.
@@ -25,7 +24,7 @@ import (
 func ListTemplates(ctx context.Context, tenant string, includePublic bool, pagination *types.Pagination) (*types.ListResponse, error) {
 	// TODO(ChenDe): Need a more efficient way to get paged items.
 	templates, err := handler.K8sClient.CycloneV1alpha1().Stages(common.TenantNamespace(tenant)).List(metav1.ListOptions{
-		LabelSelector: wfcommon.StageTemplateLabelSelector,
+		LabelSelector: common.StageTemplateLabelSelector,
 	})
 	if err != nil {
 		log.Errorf("Get templates from k8s with tenant %s error: %v", tenant, err)
@@ -35,7 +34,7 @@ func ListTemplates(ctx context.Context, tenant string, includePublic bool, pagin
 	items := templates.Items
 	if tenant != common.AdminTenant && includePublic {
 		publicTemplates, err := handler.K8sClient.CycloneV1alpha1().Stages(common.TenantNamespace(common.AdminTenant)).List(metav1.ListOptions{
-			LabelSelector: wfcommon.StageTemplateLabelSelector,
+			LabelSelector: common.StageTemplateLabelSelector,
 		})
 		if err != nil {
 			log.Errorf("Get templates from k8s with tenant %s error: %v", common.AdminTenant, err)
