+++ pkg/workflow/controller/handlers/pod/gc.go
@@ -6,8 +6,8 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	"github.com/caicloud/cyclone/pkg/common"
 	"github.com/caicloud/cyclone/pkg/k8s/clientset"
-	"github.com/caicloud/cyclone/pkg/workflow/common"
 )
 
 // IsGCPod judges whether a pod is a GC pod by check whether it has
