+++ pkg/workflow/controller/handlers/pod/operation.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/caicloud/cyclone/pkg/apis/cyclone/v1alpha1"
 	"github.com/caicloud/cyclone/pkg/k8s/clientset"
-	"github.com/caicloud/cyclone/pkg/workflow/common"
+	"github.com/caicloud/cyclone/pkg/common"
 	"github.com/caicloud/cyclone/pkg/workflow/workflowrun"
 	log "github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
